refactor: strip version off from manager adapter
